@sumeerbhola 2022-01-12
> ... The downside to this layout is that if you do need to access a historic version you incur the penalty of another Seek.

- The 2 physical sstable model is roughly how the original issue https://github.com/cockroachdb/pebble/issues/1170 was written. There are some difficulties there in that the latest version could be deleted by a higher level, so completely hiding an older version when deciding to construct an iterator is not viable. I believe what you are suggesting is exposing the older version lazily when needed, which is possibly workable. I need to think some more.
- I was trying to come up with a solution where accessing older versions would be almost the same cost as now. Doing a Seek would be much more costly.
- There are also cases (the gc ones mentioned in https://github.com/cockroachdb/pebble/issues/1170#issuecomment-958203245) that need only the size of each version (including older versions), which this implementation provides without accessing the value blocks.
- There seem to be two high-level options:
  - move older keys to a different part too: I am quite convinced that accessing these should avoid a seek. I think avoiding the seek is doable by having some notion of a continuation handle, such as (a@10=>val, a@8=>continuation handle) where the continuation handle points to all remaining key-value pairs with prefix a.
  - (the solution here) only the values are in a different part: One can still optimize NextPrefix to not do O(key-len) work (since we have the valueHandlePrefix byte) and if there are data blocks that contain only older versions one can completely skip them using block property filters (@jbowens  outlined one idea here https://github.com/cockroachdb/pebble/issues/1170#issuecomment-959329597). So the main overhead over the other option is that we would repeat the key every `BlockRestartInterval`. If there are 100s of versions in a swathe of the key space and the default value of 16 of BlockRestartInterval is insufficient we could also keep stats in the properties block about versions and use the input stats for all sstables in a compaction to choose the BlockRestartInterval.
  
  I'm still partial to the latter option since the layout and code is simpler. I'll try to add some optimized code for NextPrefix for the case of long keys and see what the cpu profiles and the block bytes tell us about the cost.