@sumeerbhola 2022-01-12
I've added a new commit with results for `NextSkipOlderVersions`. The commit has details, and here is a summary:
- With 10 versions, the results are competitive in comparison with 1 version (only 1.5x slower -- a naive iteration would be 10x slower since it would call `Next` 10 times).
- 100 versions is 10x slower (so the slowdown is accelerating since a constant slope would mean 1.5*1.5=2.25x slower). Though still better than a naive 100x slowdown, but could have been better. The major cost based on cpu profiles is inherently due to the fact that we are storing all keys in the data blocks.

I thought some more about removing older version keys out of the data blocks: there are places that do seeks using a key that has a non-empty version in CockroachDB (intent resolution; pebbleMVCCScanner using the read timestamp; deletion step of GC that seeks to the version being deleted). To support these seeks, we would need to be able to accurately position among the older versions, which adds more complexity to how these older versions are stored and indexed.

So my current thinking is that removing older version keys out of the data blocks is too risky (from a performance perspective when older versions are needed) and too complicated as a first iteration over the sstable format. @petermattis let me know what you think.