@sumeerbhola 2022-01-12
Added another commit that reduces 10 versions to 1x of 1 version and 100 versions to 6x of 1 version.