@sumeerbhola 2022-01-13
> Added another commit that reduces 10 versions to 1x of 1 version and 100 versions to 6x of 1 version.

Most of the cost (75-80%) in the 100 versions case is in decoding varints as we step through all the key versions. 
We can avoid this cost by dynamically switching (at read time) from stepping through each key to stepping through restarts. We can correctly (and cheaply) do the latter by storing a byte (bool) (at write time) with each restart key that indicates that all intervening keys since the previous restart in the same block are SETs, and the previous restart has the same prefix.